… CHANGES.md
… src/index.hpp
… src/main.cpp
… src/randstrobes.cpp
… src/randstrobes.hpp
… src/unused.cpp
… tests/baseline-commit.txt
… tests/tests.cpp