This is an extension of PR #179 (the first commits are the same). PR #179 reduced memory usage significantly when mapping using a pre-generated index, but increases memory usage at index generation time. The aim of this PR is to reduce memory usage even at index generation time.

It would have been possible to just add these commits to PR #179, but because this goes in a somewhat different direction and has different tradeoffs than what we discussed there, I’m opening a new PR.

The main downside of the method used here is that indexing time increases significantly (definitely more than 25%). I have thought about this a bit over the last weeks and am now at a point where we just need to discuss whether this is acceptable.

Memory usage is reduced to nearly the same level as when using the new index layout introduced in #179 a mapping using a pre-generated index.

Since UPPMAX servers are down for maintenance, I will need to add runtime and memory usage numbers later.

The main idea (on top of those in #179) is to [build the index directly while iterating over randstrobes](https://github.com/ksahlin/strobealign/commit/20db8150254e5adeb3e9bc5fc814aa1c8a662541). Previously, the `ind_flat_vector` was filled and then the `mers_index` and `flat_vector` were created from it. Here, we generate the  randstrobes on-the-fly and directly fill the `mers_index` and the flat vector. This is like using a generator in a Python. The previous `RandstrobeIterator` generates a vector of all randstrobes for an entire reference and uses lots of memory for large contigs. The new `RandstrobeIterator2` returns the randstrobes one by one and only keeps a small window of the relevant syncmers in memory. (In a `deque`.)

The biggest problem is that we can no longer easily estimate how much memory to pre-allocate for the hash table (`mers_index`). Previously, we could just count the distinct hash values in the sorted flat vector, which is fast and easy. Since we no longer have a full flat vector available, this is no longer possible.

If we let the hash table resize automatically without pre-allocation, memory usage grows a lot because at resize time, the old and new hash table are kept in memory so that one can be copied to the other.

For the moment, I solved this by adding a pre-processing step that generates all randstrobes without storing them and just counts how many unique hashes are created. This takes of course a lot of extra time. We also save some time because we no longer need to sort the full flat vector, only the part of it that contains randstrobes that occur more than once. Sorting time becomes nearly negligible while it was a big chunk of index generation time previously, but with all randstrobes being generated twice, runtime still increases by a bit. (I will add numbers later.)

To count unique hashes, I initially used a `robin_hood::unordered_flat_set` to keep track of all hashes, but of course also that needs a lot of memory, so I switched to using the [HyperLogLog algorithm by Flajolet et al (2007)](https://en.wikipedia.org/wiki/HyperLogLog), which uses only $O(1)$ memory.

## To Do

- [x] Add runtime and memory usage numbers
- [x] `StrobemerIndex::populate()` became quite large because I inlined `index_vector()`, `generate_and_sort_seeds()` and `randstrobes_reference()` into it. Clean it up (probably by factoring out some functions).
- [X] ~~HyperLogLog cannot deal with more than $2^{32}$ unique values, which will give problems for large references. Use HyperLogLog++ instead, which is also more accurate. (Need to find a C++ implementation somewhere.) ~~ Moved to issue #193.
- [x] `RandstrobeIterator2` gives slightly different results than `RandstrobeIterator`. Figure out why.