@ksahlin 2023-01-11
Nice! I don't have a big issue with indexing time being 25% slower, as it is already pretty fast. If the peak memory of the new implementation is kept at ~2/3 of peak memory for v0.7.1, I consider that a relatively big win.