@ksahlin 2023-01-11
Another benefit of reading the randstrobes an extra pass before indexing is for what we discussed about [64 vs 32bit hashes](https://github.com/ksahlin/strobealign/issues/27#issuecomment-1357491555). With your new approach, one could determine if it is suitable to use 32bit hashes, depending on the expected collision rate. 

For example, from memory I think indexing hg38 has about 500-600M (~2^29) unique seeds (depending on read length), perhaps this is on the border line for what is tolerated given 4.3B (2^32) unique hash values (550/4300=0.13). But for smaller genomes, it would be suitable. 

I implemented 32bit hashes (a long time ago) and saw a sizeable reduction in memory and I also think I observed a small but significant speedup (say ~10%). The speedup wouldn't be that much in your new implementation though, since you already gain some speed with storing positions directly for unique hashes. (Also, I believe any speedup in this step will not compare to changing the extension alignment to WFA).

Regardless it could be interesting to try if your new approach 'sets up' nicely for dynamically deciding the size of hash values. Not a big priority if not easily implemented I would say. What do you think?