@marcelm 2023-01-11
Here are some numbers for benchmarks run on rackham. Runtimes are extremely noisy, so take these with a grain of salt (250 could actually be 200 to 300 it appears). I will do some benchmarks on my own machine on (smaller references) overnight. That should have less noise.

But in a preliminary summary:
- Runtime goes up by a lot in some cases. It’s not as bad when we compare to the pre-pdqsort version.
- Peak memory drops a lot, but is still a little bit higher than when using a pre-generated index.

I’m still unsure whether the runtime increase isn’t too high. But dynamically deciding the size of the hash table would be a very interesting benefit of this indeed.

## Indexing time

In seconds

dataset     | before pdqsort | main (w/pdqsort) | this PR |
-|-|-|-|
rye 2x150bp | 510 | 390 | 800 |
CHM13 2x100bp | 250 | 190 | 280 |
CHM13 2x150bp | 230 | 190 | 340 |
CHM13 2x200bp | 250 | 210 | 300 |

## Peak memory when creating the index

In GiB

dataset     | main | this PR |
-|-|-|
rye 2x150bp | 47.9 | 44.7 |
CHM13 2x100bp | 28.6 | 22.6 |
CHM13 2x150bp | 28.6 | 22.4 |
CHM13 2x200bp | 28.6 | 22.3 |