@ksahlin 2023-01-11
The indexing times with v0.7.1 for CHM13 and rye were ~270 and ~600 respectively. So the numbers you display are not a drastic increase if we compare to those values (although I will wait for more stable runtimes as you mention). Peak memory for v0.7.1 was 31.5Gb and 50Gb, respectively. So a decent cut. 

> I’m still unsure whether the runtime increase isn’t too high. But dynamically deciding the size of the hash table would be a very interesting benefit of this indeed.

Well I would say that I am still positive :) Alignment time will likely be the bottleneck in many scenarios. Loosing ~30secs in indexing time for human (compared to v0.7.1) seems worth it for (i) slightly faster alignment, (ii) reduced peak memory, and (iii) setting up for 32bit hash value indexing.

(The 190 seconds for CHM with **main (w/pdqsort)** is as fast as minimap2 indexing though (191sec reported in my paper), but I don't want to become greedy.)