@marcelm 2023-01-12
Here are some more measurements.

## Indexing time

For the runtime measurements, I used [hyperfine](https://github.com/sharkdp/hyperfine), which is a dedicated benchmarking tool. I used the default settings, which means that it runs the command at least ten times (that’s where the standard deviation comes from).

dataset | 603124d (before pdqsort) | 449a2fe2f128bfa3bf196b39b590f9641f72371c (w/pdqsort) | this PR |
-|-|-|-|
*Oryza sativa* (rice, 381 Mbp) | 21.229 s ±  0.038 s |17.513 s ±  0.047 s|
*Clupea harengus* (herring, 818 Mbp) | 45.123 s ±  0.893 s|33.356 s ±  0.240 s|
*D. melanogaster* (146 Mbp) | 7.869 s ±  0.050 s | 6.518 s ±  0.027 s|
*E. coli* (4.7 Mbp) |225.2 ms ± 0.9 ms|193.4 ms ±   1.1 ms|

## Peak memory when creating the index

dataset | main (449a2fe2f128bfa3bf196b39b590f9641f72371c | this PR |
-|-|-|-|
*Oryza sativa* (rice, 381 Mbp) | 3.6 GiB | 2.8 GiB |
*Clupea harengus* (herring, 818 Mbp) | 7.0 GiB | 5.6 GiB |
*D. melanogaster* (146 Mbp) | 1.1 GiB | 0.8 GiB |
*E. coli* (4.7 Mbp) | 56.4 MiB | 64.9 MiB |