@ksahlin 2023-01-12
I guess we wait for the numbers in the last column "**this PR**" of the indexing time table. But my overall take from this is that indexing is fast and hopefully remains relatively fast with the reduced memory footprint(?).