@marcelm 2023-01-12
> I guess we wait for the numbers in the last column "this PR" of the indexing time table.

Oops, I pasted an old version of the table ... updated now.