@ksahlin 2023-01-12
It looks like counting unique strobemers is responsible for near 2x slowdown compared to w/pdqsort (see herring). Only generating randstrobes should be a quick parallel task across chromosomes. Would it be possible to parallelise over chromosomes for the two randstrobe passes? This assumes, for the first pass, that it is possible to intersect the seeds for each chromosome into a total set of unique seeds.