@marcelm 2023-01-13
The good news is that making the estimation of unique randstrobe counts multithreaded was a lot easier than I thought. The bad news is that the resulting estimates are no longer correct. They vary depending on the order in which the threads work on the contigs, and they are too large by orders of magnitude.

I’m not sure what is going on. I hope I’m just doing something wrong, but it could also be that the `HyperLogLog::merge()` method is buggy.