@marcelm 2023-01-13
> Would be interesting to see what happens for CHM13 and rye in parallell mode. Just to make sure there are no unexpected effects, such as the merging overhead overtaking sequential indexing time.

For CHM13, I tested running only the estimation part on my machine (which is possible because it only needs RAM for the reference). Runtime with 1 thread was 84 seconds and 23 seconds with 4 threads. 84/23=3.7, which is quite ok considering that other things were running on my machine at the same time.

Merging is quite fast (it’s essentially a pairwise `std::max` on two relatively short vectors), and it is only done once for each contig (plus once for each thread). CHM13 only 25 contigs, so there is virtually no overhead. Herring and fruit fly on the other hand, are good test cases because they have 6914 and 1870 contigs, respectively.

> Overall I am very positive about this approach and pull request. What is your opinion @marcelm ?

I’ve been working on this and thinking about this for way too long and would like to be able to finally work on something else. Until now, I wasn’t sure whether this approach was viable, but with this most recent change, I’m ok with the slowdown.

I’m still not fully sure where the slowdown comes from. Here is an example for herring:

Before (main):
* Generating randstrobes: 14 s
* Sorting: 6 s
* Generating hash table: 11 s
* Other stuff: 1 s
* Total: 32 s

After:
* Time estimating number of unique hashes: 17 s (or faster with more than one thread)
* Generating randstrobes and filling the hash table at the same time: 40 s
* Sorting: 1 s

I don’t get where the 40 seconds come from – it should only take a little bit longer than 25 seconds.

A couple of seconds are lost because the `RandstrobeIterator2` that generates the strobemers on-the-fly is a little bit slower than `RandstrobeIterator`. It needs ~16 seconds instead of 14. (I have tried to find out why but cannot say. But I don’t suggest using it for the reads for that reason.)

When filling the hash table on-the-fly (20db815), an additional lookup is made for each randstrobe in order to determine whether it already exists in the table (and if it does, then some extra work is necessary, but that happens only for non-unique randstrobe and therefore not often), but that extra lookup seems to be cost only about 2 seconds – I don’t know where the rest of the time goes.

> are you waiting to merge until the TODO's you added are completed or is there anything else to consider?

I would like to ignore the To Do about switching to HyperLogLog++ for the moment. Regarding different results for the RandstrobeIterator2, I think there was somehting about how randstrobes were generated at the end of a sequence. Accuracy changes a little bit; I should probably investigate:
* rye 2x150 90.22765 → 90.21695
* CHM13 2x100 93.21825 → 93.2193
* CHM13 2x200 94.4156 → 94.4089
* CHM13 2x150 94.1543 → 94.145